refactor: add separate errors file
